requirements.txt
numpy>=1.21.0
pandas>=1.3.0
networkx>=2.6.0
sentence-transformers>=2.2.0
torch>=1.9.0
openai>=1.0.0
pyyaml>=6.0
scikit-learn>=1.0.0

